# openapi/analytics-admin.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "metrics": { "bearerFormat": "JWT", "scheme": "bearer", "type": "http" },
      "session": {
        "in": "cookie",
        "name": "__Host-analytics_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Tenant-safe operational and sales ingestion API",
    "title": "Analytics Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/data/inventory": {
      "post": {
        "operationId": "DataController_ingestInventory_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "location": {
                    "additionalProperties": false,
                    "properties": {
                      "countryCode": {
                        "maxLength": 2,
                        "minLength": 2,
                        "pattern": "^[A-Z]{2}$",
                        "type": "string"
                      },
                      "externalId": {
                        "maxLength": 160,
                        "minLength": 1,
                        "type": "string"
                      },
                      "kind": {
                        "enum": ["warehouse", "store", "virtual", "other"],
                        "type": "string"
                      },
                      "name": {
                        "maxLength": 120,
                        "minLength": 1,
                        "type": "string"
                      },
                      "timezone": {
                        "maxLength": 80,
                        "minLength": 1,
                        "type": "string"
                      }
                    },
                    "required": ["externalId", "name"],
                    "type": "object"
                  },
                  "onHandQuantity": { "minimum": 0, "type": "integer" },
                  "product": {
                    "additionalProperties": false,
                    "properties": {
                      "externalId": {
                        "maxLength": 160,
                        "minLength": 1,
                        "type": "string"
                      },
                      "name": {
                        "maxLength": 200,
                        "minLength": 1,
                        "type": "string"
                      },
                      "sku": {
                        "maxLength": 160,
                        "minLength": 1,
                        "type": "string"
                      }
                    },
                    "required": ["externalId", "name", "sku"],
                    "type": "object"
                  },
                  "reorderPoint": { "minimum": 0, "type": "integer" },
                  "reservedQuantity": { "minimum": 0, "type": "integer" },
                  "sourceUpdatedAt": { "format": "date-time", "type": "string" }
                },
                "required": [
                  "location",
                  "onHandQuantity",
                  "product",
                  "reorderPoint",
                  "sourceUpdatedAt"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "locationId": { "format": "uuid", "type": "string" },
                    "productId": { "format": "uuid", "type": "string" },
                    "status": { "enum": ["accepted"], "type": "string" }
                  },
                  "required": ["locationId", "productId", "status"],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "summary": "Create or replace one inventory position snapshot",
        "tags": ["Data ingestion"]
      }
    },
    "/api/v1/organizations/{organizationId}/data/orders": {
      "post": {
        "operationId": "DataController_ingestOrder_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "channel": {
                    "additionalProperties": false,
                    "properties": {
                      "externalId": {
                        "maxLength": 160,
                        "minLength": 1,
                        "type": "string"
                      },
                      "kind": {
                        "enum": [
                          "marketplace",
                          "storefront",
                          "pos",
                          "wholesale",
                          "other"
                        ],
                        "type": "string"
                      },
                      "name": {
                        "maxLength": 120,
                        "minLength": 1,
                        "type": "string"
                      }
                    },
                    "required": ["externalId", "kind", "name"],
                    "type": "object"
                  },
                  "currency": {
                    "maxLength": 3,
                    "minLength": 3,
                    "pattern": "^[A-Z]{3}$",
                    "type": "string"
                  },
                  "customerReference": {
                    "additionalProperties": false,
                    "properties": {
                      "countryCode": {
                        "maxLength": 2,
                        "minLength": 2,
                        "pattern": "^[A-Z]{2}$",
                        "type": "string"
                      },
                      "externalId": {
                        "maxLength": 160,
                        "minLength": 1,
                        "type": "string"
                      },
                      "metadata": {
                        "additionalProperties": true,
                        "type": "object"
                      }
                    },
                    "required": ["externalId"],
                    "type": "object"
                  },
                  "discountMinor": { "minimum": 0, "type": "integer" },
                  "externalId": {
                    "maxLength": 160,
                    "minLength": 1,
                    "type": "string"
                  },
                  "fulfilments": {
                    "items": {
                      "additionalProperties": false,
                      "properties": {
                        "carrier": { "type": "string" },
                        "deliveredAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "externalId": {
                          "maxLength": 160,
                          "minLength": 1,
                          "type": "string"
                        },
                        "locationExternalId": {
                          "maxLength": 160,
                          "minLength": 1,
                          "type": "string"
                        },
                        "shippedAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                          ],
                          "type": "string"
                        },
                        "trackingReference": {
                          "maxLength": 160,
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": ["externalId", "status"],
                      "type": "object"
                    },
                    "maxItems": 100,
                    "type": "array"
                  },
                  "items": {
                    "items": {
                      "additionalProperties": false,
                      "properties": {
                        "discountMinor": { "minimum": 0, "type": "integer" },
                        "externalId": {
                          "maxLength": 160,
                          "minLength": 1,
                          "type": "string"
                        },
                        "name": { "minLength": 1, "type": "string" },
                        "productExternalId": {
                          "maxLength": 160,
                          "minLength": 1,
                          "type": "string"
                        },
                        "quantity": {
                          "maximum": 1000000,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "sku": {
                          "maxLength": 160,
                          "minLength": 1,
                          "type": "string"
                        },
                        "taxMinor": { "minimum": 0, "type": "integer" },
                        "totalMinor": { "minimum": 0, "type": "integer" },
                        "unitPriceMinor": { "minimum": 0, "type": "integer" }
                      },
                      "required": [
                        "externalId",
                        "productExternalId",
                        "sku",
                        "name",
                        "quantity",
                        "unitPriceMinor",
                        "totalMinor"
                      ],
                      "type": "object"
                    },
                    "minItems": 1,
                    "type": "array"
                  },
                  "location": {
                    "additionalProperties": false,
                    "properties": {
                      "countryCode": {
                        "maxLength": 2,
                        "minLength": 2,
                        "pattern": "^[A-Z]{2}$",
                        "type": "string"
                      },
                      "externalId": {
                        "maxLength": 160,
                        "minLength": 1,
                        "type": "string"
                      },
                      "kind": {
                        "enum": ["warehouse", "store", "virtual", "other"],
                        "type": "string"
                      },
                      "name": {
                        "maxLength": 120,
                        "minLength": 1,
                        "type": "string"
                      },
                      "timezone": {
                        "maxLength": 80,
                        "minLength": 1,
                        "type": "string"
                      }
                    },
                    "required": ["externalId", "name"],
                    "type": "object"
                  },
                  "metadata": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "occurredAt": { "format": "date-time", "type": "string" },
                  "orderNumber": { "minLength": 1, "type": "string" },
                  "payment": {
                    "additionalProperties": false,
                    "properties": {
                      "authorizedMinor": { "minimum": 0, "type": "integer" },
                      "capturedMinor": { "minimum": 0, "type": "integer" },
                      "refundedMinor": { "minimum": 0, "type": "integer" },
                      "status": {
                        "enum": [
                          "pending",
                          "authorized",
                          "captured",
                          "partially_refunded",
                          "refunded",
                          "failed"
                        ],
                        "type": "string"
                      }
                    },
                    "required": ["status"],
                    "type": "object"
                  },
                  "returns": {
                    "items": {
                      "additionalProperties": false,
                      "properties": {
                        "amountMinor": { "minimum": 0, "type": "integer" },
                        "externalId": {
                          "maxLength": 160,
                          "minLength": 1,
                          "type": "string"
                        },
                        "receivedAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "requestedAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "requested",
                            "approved",
                            "received",
                            "rejected",
                            "refunded"
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "amountMinor",
                        "externalId",
                        "requestedAt",
                        "status"
                      ],
                      "type": "object"
                    },
                    "maxItems": 100,
                    "type": "array"
                  },
                  "shippingMinor": { "minimum": 0, "type": "integer" },
                  "sourceUpdatedAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "status": {
                    "enum": [
                      "pending",
                      "confirmed",
                      "fulfilled",
                      "cancelled",
                      "refunded"
                    ],
                    "type": "string"
                  },
                  "subtotalMinor": { "minimum": 0, "type": "integer" },
                  "taxMinor": { "minimum": 0, "type": "integer" },
                  "totalMinor": { "minimum": 0, "type": "integer" }
                },
                "required": [
                  "channel",
                  "currency",
                  "externalId",
                  "items",
                  "occurredAt",
                  "orderNumber",
                  "status",
                  "subtotalMinor",
                  "totalMinor"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "externalId": { "type": "string" },
                    "orderId": { "format": "uuid", "type": "string" },
                    "status": { "enum": ["accepted"], "type": "string" }
                  },
                  "required": ["externalId", "orderId", "status"],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "summary": "Create or replace a source order snapshot",
        "tags": ["Data ingestion"]
      }
    },
...
  },
  "servers": [],
...
}